Refactor LV-Lite and DM algorithms for clarity and consistency
Renamed several classes and methods to improve clarity and align with naming conventions (e.g., LvLiteDm → DmLvLite, DM → DmPc). Added new fields and getter methods to Lv
